bft_interp: Use load/as u8 in trait
Change the implementation to match the homework review at the start of
the session 06 stream.

Q: Why is it ok that we put to/from u8 methods in the generic cell-kind
trait?

I think it's probably because Brainfuck expects data in/out to be in u8.
Maybe this will become more clear when we change the type of
BrainfuckCellKind.